requirements.txt
amqp==5.2.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.0
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.1
cssselect2==0.7.0
datetime-truncate==1.1.1
Django==5.0
django-admin-charts==1.4.0
django-bower==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-daraja==1.3.0
django-grappelli==4.0.1
django-jazzmin==3.0.0
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nested-admin==4.1.1
django-nvd3==0.10.1
django-redis==5.4.0
django-suit==0.2.28
django-timezone-field==7.0
django-unfold==0.39.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==28.4.1
fonttools==4.54.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
html5lib==1.1
httplib2==0.22.0
idna==3.8
importlib_metadata==8.5.0
Jinja2==3.1.4
kombu==5.4.1
lxml==5.3.0
MarkupSafe==2.1.5
oauth2client==1.5.2
oscrypto==1.3.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyee==11.1.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.1.4
pypdf==5.0.0
pyphen==0.16.0
pypng==0.20220715.0
pyppeteer==2.0.0
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
xhtml2pdf==0.2.16
zipp==3.20.2
zopfli==0.2.3
